requirements.txt
fastapi
uvicorn
pillow
numpy
scikit-image
scikit-learn
